Properly reject "link proto" for DLT_BACNET_MS_TP.
When the BACnet MS/TP case in gen_linktype() tests for the arbitrary
number 0x55FF0000 using the bitmask 0xFFFF0000, the only obvious
mismatch between the problem space and the solution space is the
discarding of the low 16 bits in a 32-bit value instead of using a plain
16-bit comparison.

However, the main problem is that the bytes 0x55 0xFF comprise the MS/TP
frame preamble and gen_linktype() matches any well-formed MS/TP frame
regardless of the actual contents.  This means any explicit or implicit
tests for a link-layer protocol ID always match, for example, "port 80"
seemingly works, but erroneously interprets MS/TP frames as IPv4 and
IPv6 packets.

Remove the case from gen_linktype() to have pcap_compile() reject input
that is semantically invalid in the first place.  Add a comment and
update the tests.