Fix GENEVE- and VXLAN-encapsulated matching.
In my recent commit 1687065 the filter programs for "vxlan [NUM]",
"vxlan and vxlan", "geneve [NUM]" and "geneve and geneve" have become
substantially shorter.  As it turns out, the missing instructions are
the outputs of gen_geneve_offsets() and gen_vxlan_offsets(), so
filtering of protocols encapsulated in GENEVE and VXLAN is no longer
correct.  Because libpcap does not have any apply filter tests for the
affected syntax, this was not apparent until a tcpdump test failed.

The instructions went missing because gen_geneve() and gen_vxlan() need
to use these as a block, but instead of making a new block from scratch
the functions take the result of gen_true() and modify it to add side
effects.  After the commit the block with the wanted side effects ended
up incorrectly marked as a Boolean constant and susequently gen_and()
started to discard it.

In gen_cond() take an optional list of side effect instructions and make
a Boolean constant block a specific case of a block with a degenerate
branch statement.  Switch gen_geneve() and gen_vxlan() from gen_true()
to gen_uncond() to reduce the violation of layers.  Restore the six
previously affected tests exactly as they were before.