Improve Boolean de-constification a bit more.
It is a valid use case to prepend items to the list of side effect
statements of a block that already exists and/or is not a Boolean
constant, so instead of mixing two unrelated things in gen_uncond()
revert commit d0d58d9 and add a new function, sprepend_to_block(), to do
the prepending part and to mind Boolean constants.  Use it in
gen_geneve() and gen_vxlan(), which have already proved to require this
solution.

Also use the new function in insert_compute_vloffsets(), the result of
which is an argument to gen_and() in finish_parse() for DLT_PPI only, in
which case its block argument in the current implementation is never a
Boolean constant, but let's just make it obviously correct anyway.
Ibid., clarify that the passing of a block statements to
gen_load_802_11_header_len() does not require an immediate Boolean
de-constification of the block.

Also use it in gen_vlan_patch_tpid_test() and gen_vlan_patch_vid_test(),
which take the block argument from gen_vlan_tpid_test() and
gen_vlan_vid_test() respectively, therefore the argument is never a
Boolean constant.  But let's not rely on that either.