rpcapd: Get port size check right in fileconf_read(). [skip ci]
The block that fills the port member of an array element takes sizeof()
of the address member, which is a bug, but a benign one because address
and port are of the same exact size.  Fix it before it gets worse.

Reported by Aniruddhan Murali and Noble Saji Mathews, both of University
of Waterloo.  See GH #1566.