filtertest: Fix two NULL pointer dereferences. [skip ci]
TESTrun among other accept tests runs filtertest with an empty filter
string, but it does not run filtertest without a filter string, which is
exactly what my commit a64b807 has broken, as Francois-Xavier notes:

$ filtertest EN10MB
Segmentation fault

Do not substitute NULL with an empty string because that would get in
the way of testing the NULL filter later if required.  Instead add a
missing check for NULL in the new code block and another in the old code
block that prints the expression if optimizer debugging is enabled.  Add
more comments.