Enable various primitives for DLT_DSA_TAG_DSA.
This works identical to DLT_DSA_TAG_BRCM just before because none of the
primitives are DLT-specific and the infix header offset and size are the
same.  dsa.pcap comes from tcpdump tests.  Place the new VLAN and MPLS
accept tests a bit further away from the existing VLAN and MPLS tests
because otherwise git-diff fails to recognize new lines as new lines.