Add inbound/outbound support for two DSA DLTs.
In gen_inbound_outbound() add cases for DLT_DSA_TAG_BRCM and
DLT_DSA_TAG_DSA.  Cover the new code paths with accept and filter
tests.  For the latter use the existing savefiles.  Annotate the
contents of dsa.pcap better to make it easier to interpret the results.
Make the definition of inbound/outbound in pcap-filter(7) a bit more up
to date.

In my tests the new code handles DLT_DSA_TAG_BRCM correctly for a live
capture as well.