FreeBSD USB: remove interface creation for FreeBSD 16.0
In FreeBSD 16.0 the USB bpf(4) taps are always available and the creation
of pseudo-interfaces, that used to exist merely for a bpf(4) tap, is
neither required nor is possible.  The list of all tap points can be
obtained with BIOCGETIFLIST.  Note that we still use
pcapint_findalldevs_interfaces() as the main method of collecting bpf
taps, and we use BIOCGETIFLIST only to fill the gaps there - get those bpf
taps, that are not a network interface.