Fix an OOBR in pcap_findalldevs_ex().
When the PCAP_SRC_FILE case code enforces a trailing slash in a
directory name, it does not handle an edge case correctly: when the
source string is "file://", the name is an empty string and the code
reads one byte from an existing, but unrelated address in stack memory,
which can erroneously convert the source string to "file:///" before
further processing.

This bug was originally reported by GPTFuzz and subsequently by Jin Wei,
Kunwei Qian and Ping Chen.