CVE-2025-11964: Complete the change log entry.
Add the CVE ID only: the fix has already been made in the earlier commit
aebfca1.  Guy Harris has found that this problem is reproducible only on
Windows, and only under very specific circumstances that are not
impossible in principle, but are so unlikely that it is not clear if a
vulnerable environment actually exists.

For posterity, this bug was reported by Aniruddhan Murali, Noble Saji
Mathews, Mahmoud Alfadel, Mei Nagappan and Meng Xu, all from University
of Waterloo and all contributors of the technology FuzzSlice that helped
detect this bug.