Introduce a routine to parse decimal numbers in a string.
This wraps strtoul() to do more error checking.

It's currently used only in one place, but there are other places where
it could be useful.

Make translatetest's getopt() code do the standard "add optind to argv
and subtract it frm argc" after the loop, so the first argument after
the option arguments is argv[0] and there are a total of argc post-options
arguments, so that it can handle a "--" argument to allow "-1" to be
passed as a non-option argument, even with GNU getopt().

Add some tests, courtesy of Denis Ovsienko.